# app/Http/Controllers/CompanyController.php
<?php

namespace App\Http\Controllers;

use App\Http\Requests\CreateCompanyRequest;
use App\Http\Requests\UpdateCompanyRequest;
use App\Http\Resources\CompanyResource;
use App\Models\Company;
use App\Services\CompanyService;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Symfony\Component\HttpFoundation\Response as ResponseAlias;


class CompanyController extends Controller
{
    /**
     * @var CompanyService
     */
    protected $service;
    /**
     * @var
     */
    protected $model;

    /**
     * @param CompanyService $companyService
     */
    public function __construct(CompanyService $companyService)
    {
        $this->service = $companyService;
    }

    /**
     * @OA\Get(
     *      path="/companies",
     *      tags={"Company(Creditors)"},
     *      summary="Get list of projects",
     *      description="Returns list of projects",
     *      security={ * {"sanctum": {}}, * },
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation"
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *     )
     */
    public function index(): JsonResponse
    {

        try {
            $this->model = $this->service->index();
            return $this->respondWithResource($this->model);
        } catch (\Exception $e) {
            return $this->respondError($e->getMessage());
        }
    }

    /**
     * @OA\Post(
     *      path="/companies",
     *      tags={"Company(Creditors)"},
     *      summary="Get list of projects",
     *      description="Returns list of projects",
     *      security={ * {"sanctum": {}}, * },
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation"
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *     )
     */
    public function store(CreateCompanyRequest $request): JsonResponse
    {
        try {
            $this->model = $this->service->create($request);
            return $this->respondWithResource(new CompanyResource ($this->model), '', ResponseAlias::HTTP_CREATED);
        } catch (\Exception $e) {
            return $this->respondError($e->getMessage());
        }

    }

    /**
     * @OA\Get(
     *      path="/companies/{id}",
     *      tags={"Company(Creditors)"},
     *      summary="Get company information",
     *      description="Returns company data",
     *      @OA\Parameter(
     *          name="id",
     *          description="Project id",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *      ),
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\JsonContent(
     *       @OA\Property(property="success", type="boolean", example="false")
     *        ),
     *       @OA\Property(property="message", type="string", example="The given data was invalid.")
     *        ),
     *      @OA\Property(property="data", type="array", example="[]")
     *        )
     * )
     *       ),
     *      @OA\Response(
     *          response=400,
     *          description="Bad Request"
     *      ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     * )
     */
    public function show(Company $company): JsonResponse
    {
        try {
            return $this->respondWithResource(new CompanyResource($company));
        } catch (\Exception $e) {
            return $this->respondError($e->getMessage());
        }
    }


    /**
     * @OA\Put(
     *      path="/companies",
     *      tags={"Company(Creditors)"},
     *      summary="Get list of projects",
     *      description="Returns list of projects",
     *      security={ * {"sanctum": {}}, * },
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation"
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *     )
     */
    public function update(UpdateCompanyRequest $request): JsonResponse
    {
        try {
            $this->model = $this->service->update($request);
            return $this->respondWithResource((new CompanyResource($this->model)));
        } catch (\Exception $e) {
            return $this->respondError($e->getMessage());
        }
    }

    /**
     * @OA\Delete (
     *      path="/companies",
     *      tags={"Company(Creditors)"},
     *      summary="Get list of projects",
     *      description="Returns list of projects",
     *      security={ * {"sanctum": {}}, * },
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation"
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      )
     *     )
     */
    public function destroy(Request $request, Company $company): JsonResponse
    {
        //
        try {
            $company->delete();
            return $this->respondSuccess('Company Deleted successfully');
        } catch (\Exception $e) {
            return $this->respondError($e->getMessage());
        }

    }
}